use generics
git-svn-id: https://svn.java.net/svn/jide-oss~svn/trunk@143 8ffc590b-27e5-ea7f-97f6-895640bc1b3e

+++ src/com/jidesoft/grouper/date/DateDayOfMonthGrouper.java
@@ -27,8 +27,8 @@ public class DateDayOfMonthGrouper extends DateGrouper {
 
     public Object getValue(Object value) {
         Object field = getCalendarField(value, Calendar.DAY_OF_MONTH);
-        if (field instanceof Integer && ((Integer) field).intValue() >= 1 && ((Integer) field).intValue() <= getAvailableGroups().length) {
-            return getAvailableGroups()[((Integer) field).intValue() - 1];
+        if (field instanceof Integer && (Integer) field >= 1 && (Integer) field <= getAvailableGroups().length) {
+            return getAvailableGroups()[(Integer) field - 1];
         }
         else {
             return null;
