+++ src/com/jidesoft/grouper/date/DateDayOfWeekInMonthGrouper.java
@@ -32,8 +32,8 @@ public class DateDayOfWeekInMonthGrouper extends DateGrouper {
 
     public Object getValue(Object value) {
         Object field = getCalendarField(value, Calendar.DAY_OF_WEEK_IN_MONTH);
-        if (field instanceof Integer && ((Integer) field).intValue() >= 0 && ((Integer) field).intValue() < getAvailableGroups().length) {
-            return getAvailableGroups()[((Integer) field).intValue()];
+        if (field instanceof Integer && (Integer) field >= 0 && (Integer) field < getAvailableGroups().length) {
+            return getAvailableGroups()[((Integer) field)];
         }
         else {
             return null;
