+++ src/com/jidesoft/grouper/date/DateWeekOfMonthGrouper.java
@@ -27,8 +27,8 @@ public class DateWeekOfMonthGrouper extends DateGrouper {
 
     public Object getValue(Object value) {
         Object field = getCalendarField(value, Calendar.WEEK_OF_MONTH);
-        if (field instanceof Integer && ((Integer) field).intValue() >= 0 && ((Integer) field).intValue() < getAvailableGroups().length) {
-            return getAvailableGroups()[((Integer) field).intValue()];
+        if (field instanceof Integer && (Integer) field >= 0 && (Integer) field < getAvailableGroups().length) {
+            return getAvailableGroups()[((Integer) field)];
         }
         else {
             return null;
